Semi-working using the new ConversationHandler
